Avoid casting warnings

+++ src/isct/triangle.c
@@ -1422,10 +1422,10 @@ REAL area;                                      /* The area of the triangle. */
 /**                                                                         **/
 
 #ifdef ANSI_DECLARATORS
-void triexit(INT status)
+void triexit(int status)
 #else /* not ANSI_DECLARATORS */
 void triexit(status)
-INT status;
+int status;
 #endif /* not ANSI_DECLARATORS */
 
 {
@@ -6964,7 +6964,7 @@ struct badtriang *badtri;
     length *= multiplier;
   }
   /* `length' is approximately squareroot(2.0) to what exponent? */
-  exponent = 2.0 * exponent + (length > SQUAREROOTTWO);
+  exponent = (INT)(2.0 * exponent + (length > SQUAREROOTTWO));
   /* `exponent' is now in the range 0...2047 for IEEE double precision.   */
   /*   Choose a queue in the range 0...4095.  The shortest edges have the */
   /*   highest priority (queue 4095).                                     */
